Rename to GoogleSearchStore
