+++ src/apps/settings/SettingsModal.tsx
@@ -6,7 +6,7 @@ import BuildCircleIcon from '@mui/icons-material/BuildCircle';
 import { ElevenlabsSettings } from '~/modules/elevenlabs/ElevenlabsSettings';
 import { GoodModal } from '~/common/components/GoodModal';
 import { ProdiaSettings } from '~/modules/prodia/ProdiaSettings';
-import { SearchSettings } from '~/modules/search/SearchSettings';
+import { SearchSettings } from '~/modules/google/SearchSettings';
 import { useUIStateStore } from '~/common/state/store-ui';
 
 import { UISettings } from './UISettings';
