+++ src/modules/google/search.types.ts
